myproject/requirements.txt
django==5.0.1
django-db==0.0.7
djangorestframework==3.14.0
matplotlib==3.5.1
numpy==1.22.2
Pillow==9.0.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0